Move DispatchWorkloop to a separate file
